feat: add entity hierarchy

+++ src/com/github/artemkoloshva/model/entity/Predator.java
@@ -1,9 +1,9 @@
 package com.github.artemkoloshva.model.entity;
 
 public abstract class Predator extends Creature implements Attacker {
-    private final int damage;
+    private final double damage;
 
-    public Predator(int healthPoints, int speed, int damage) {
+    public Predator(int healthPoints, int speed, double damage) {
         super(healthPoints, speed);
         this.damage = damage;
     }
